Move defaultNames and EventColor to base Start Event
