+++ src/components/nodes/baseStartEvent/eventShape.js
