+++ src/components/nodes/defaultNames.js
@@ -1,4 +1,4 @@
-export { default as defaultStartNames } from '@/components/nodes/startEvent/defaultNames';
+export { default as defaultStartNames } from '@/components/nodes/baseStartEvent/defaultNames';
 export { default as defaultEndNames } from '@/components/nodes/endEvent/defaultNames';
 export { default as defaultTaskNames } from '@/components/nodes/task/defaultNames';
 export { default as defaultGatewayNames } from '@/components/nodes/gateway/defaultNames';
