+++ src/components/nodes/signalStartEvent/index.js
@@ -2,8 +2,8 @@ import component from './signalStartEvent.vue';
 import merge from 'lodash/merge';
 import cloneDeep from 'lodash/cloneDeep';
 import startEventConfig from '../startEvent';
-import { signalSelector, default as signalEventDefinition } from '../signalEventDefinition';
-import defaultNames from '@/components/nodes/startEvent/defaultNames';
+import { default as signalEventDefinition, signalSelector } from '../signalEventDefinition';
+import defaultNames from '@/components/nodes/baseStartEvent/defaultNames';
 
 const id = 'processmaker-modeler-signal-start-event';
